Completed 01 exercise
Edited css to match the desired result. Key learning includes acknowledging that
div's are block elements by default and setting a margin-left to auto will result
in the block to sit as far right of the line. This is due to the right margin not
being specified and defaulting to 0 while auto causes left margin to take all
specified space (200px).